+++ resources/views/blog/index.blade.php
@@ -60,4 +60,15 @@
             </div>
         </div>
     </section>
+
+    {{-- Blog Index Schema --}}
+    @push('schema')
+    @php
+        $schemaService = app(\App\Services\Schema\SchemaService::class);
+    @endphp
+    {!! \App\Services\Schema\SchemaService::toScript($schemaService->breadcrumbs([
+        ['label' => 'Trang chủ', 'url' => route('home')],
+        ['label' => 'Blog'],
+    ])) !!}
+    @endpush
 </x-layouts.app>